Add ability to define cookies used for site
CIVIC control has an `optionalCookies` property that contains different
cookies used by the site. When editing the CIVIC cookie control
settings the content editor can manage a collection of cookies,
replicating a `BagPart`, that contains content types with a sterotype
of "Cookie". Currently there is only a `RawCookie`, which contains all
the properties for what the CIVIC control is expecting within
`optionalCookies`.

In order to convert a cookie content item in to a JSON representation
of an optional cookie there is a class that implements `ICookieTypw`
and defines which content type it can convert. Plan is to have different
content types to represent common cookies (e.g. Google Analytics).